+++ rtiow/V/A14.v
@@ -477,9 +477,9 @@ $if dbg? {
 			col = vec.div(col, ns)
 			// Gamma 2 correction (square root)
 			col = vec.Vec3{math.sqrtf(col.x),math.sqrtf(col.y),math.sqrtf(col.z)}
-			ir := int(255.99 * col.x)
-			ig := int(255.99 * col.y)
-			ib := int(255.99 * col.z)
+			ir := int(f32(255.99) * col.x)
+			ig := int(f32(255.99) * col.y)
+			ib := int(f32(255.99) * col.z)
 			if fnameout != '' {
 				bytes[((ny -1 - j) * nx + i) * 3 + 0] = ir
 				bytes[((ny -1 - j) * nx + i) * 3 + 1] = ig
